+++ src/styles/index.js
@@ -18,6 +18,6 @@ import "./buttons/IrisToggle.css";
 
 import "./cards/BookCard.css";
 import "./cards/ComingSoon.css";
-import "./calibration/CalibrationConsent.css";
+import "./cards/CalibrationCard.css";
 import "./calibration/CalibrationOverlay.css";
 import "./calibration/CalibrationTarget.css";